+++ src/Makefile.am
@@ -203,6 +203,7 @@ BITCOIN_CORE_H = \
   blsct/wallet/keyman.h \
   blsct/wallet/keyring.h \
   blsct/wallet/txfactory.h \
+  blsct/wallet/txfactory_global.h \
   blsct/wallet/verification.h \
   dandelion.h \
   chain.h \
@@ -630,6 +631,7 @@ libbitcoin_wallet_a_SOURCES = \
   blsct/wallet/keyman.cpp \
   blsct/wallet/keyring.cpp \
   blsct/wallet/txfactory.cpp \
+  blsct/wallet/txfactory_global.cpp \
   blsct/wallet/verification.cpp \
   wallet/coincontrol.cpp \
   wallet/context.cpp \
