+++ src/Makefile.test_util.include
@@ -55,7 +55,7 @@ libtest_util_a_SOURCES = \
   blsct/public_key.cpp \
   blsct/public_keys.cpp \
   blsct/signature.h \
-  blsct/wallet/txfactory.cpp \
+  blsct/wallet/txfactory_global.cpp \
   test/util/blockfilter.cpp \
   test/util/coins.cpp \
   test/util/json.cpp \
