+++ src/test/blsct/wallet/validation_tests.cpp
@@ -35,7 +35,7 @@ BOOST_FIXTURE_TEST_CASE(validation_test, TestingSetup)
     COutPoint outpoint{txid, /*nIn=*/0};
 
     Coin coin;
-    auto out = blsct::TxFactory::CreateOutput(recvAddress, 1000 * COIN, "test");
+    auto out = blsct::CreateOutput(recvAddress, 1000 * COIN, "test");
     coin.nHeight = 1;
     coin.out = out.out;
 
@@ -67,7 +67,7 @@ BOOST_FIXTURE_TEST_CASE(validation_reward_test, TestingSetup)
     CMutableTransaction tx;
     std::string dummyAddress = "xNVGDQwqMnD7Wi13X5FutkC11VwAf9oaZuDMs1sXUPHgGjmkh5UukhJihW56uTY4GMMLEfzF536N3zMAzMcWA6mxfm6sjsQ9k49cUNBN1B6xAGqGE2bnBsBPijYsMMQNuQjrRDLn5rs";
 
-    auto out = blsct::TxFactory::CreateOutput(dummyAddress, 900 * COIN, " Reward ");
+    auto out = blsct::CreateOutput(dummyAddress, 900 * COIN, " Reward ");
     tx.vout.push_back(out.out);
     tx.txSig = out.GetSignature();
